Refactor application singleton handling in Myster: replace anonymous listener with SingleInstanceLaunchHandler for improved clarity and functionality; enhance relaunch logic and ensure GUI readiness.

+++ src/main/java/com/myster/search/ui/ClientGenericHandleObject.java
@@ -87,7 +87,7 @@ public class ClientGenericHandleObject implements FileTypeColumnHandler {
                 String size = result.getMetaData(keyarray[1]);
                 try {
                     if (size != null && sortableSize == null) {
-                        sortableSize = new SortableByte(Integer.parseInt(size));
+                        sortableSize = new SortableByte(Long.parseLong(size));
                     }
                     return (size == null ? NOT_IN : sortableSize);
                 } catch (NumberFormatException ex) {
